Render Properties panel in place of the original frontmatter block
Previously the properties panel was always prepended to the top of the
rendered HTML, even if the frontmatter block itself appeared later in
the source (e.g. after a leading heading). Replace the strip-then-prepend
approach with an in-place splice, using the same start/end indices to
substitute the rendered properties HTML directly where the raw YAML
block was, preserving document order.
